+++ src/main/java/calculator/PriceMovingAverage.java
@@ -10,10 +10,10 @@ import util.MyUtils;
 
 public class PriceMovingAverage {
 
-  private final HashMap<Integer, HashMap<String, Float>> m_EMAMap;
-  private final HashMap<Integer, HashMap<String, Float>> m_SMAMap;
+  private final HashMap<Integer, HashMap<String, Double>> m_EMAMap;
+  private final HashMap<Integer, HashMap<String, Double>> m_SMAMap;
   //Standard deviation
-  private final HashMap<Integer, HashMap<String, Float>> m_STDMap;
+  private final HashMap<Integer, HashMap<String, Double>> m_STDMap;
 
   public PriceMovingAverage() {
     m_EMAMap = new HashMap();
@@ -21,25 +21,25 @@ public class PriceMovingAverage {
     m_STDMap = new HashMap();
   }
 
-  public Float getEMA(String date, int emaDuration,
+  public Double getEMA(String date, int emaDuration,
           ConcurrentHashMap<String, Object[]> rawDataMap) throws ParseException {
 
     Object[] raw_data = rawDataMap.get(date);
     if (raw_data == null) {
       return null;
     }
-    float p = (float) raw_data[YAHOO_DATA_INDEX.CLOSE.ordinal()];
-    Float ema = getRawEMA(date, emaDuration, rawDataMap);
+    double p = (double) raw_data[YAHOO_DATA_INDEX.CLOSE.ordinal()];
+    Double ema = getRawEMA(date, emaDuration, rawDataMap);
     if (ema == null) {
       return ema;
     } else {
-      return p / (float) ema;
+      return p / (double) ema;
     }
   }
 
-  public Float getRawEMA(String date, int emaDuration,
+  public Double getRawEMA(String date, int emaDuration,
           ConcurrentHashMap<String, Object[]> rawDataMap) throws ParseException {
-    HashMap<String, Float> ema = m_EMAMap.get(emaDuration);
+    HashMap<String, Double> ema = m_EMAMap.get(emaDuration);
     if (ema == null) {
       m_EMAMap.put(emaDuration, new HashMap());
       ema = m_EMAMap.get(emaDuration);
@@ -55,18 +55,18 @@ public class PriceMovingAverage {
     if (raw_data == null) {
       return null;
     }
-    float p = (float) raw_data[YAHOO_DATA_INDEX.CLOSE.ordinal()];
-    Float sma = getRawSMA(date, smaDuration, rawDataMap);
+    double p = (double) raw_data[YAHOO_DATA_INDEX.CLOSE.ordinal()];
+    Double sma = getRawSMA(date, smaDuration, rawDataMap);
     if (sma == null) {
       return sma;
     } else {
-      return p / (float) sma;
+      return p / (double) sma;
     }
   }
 
-  public Float getRawSMA(String date, int smaDuration,
+  public Double getRawSMA(String date, int smaDuration,
           ConcurrentHashMap<String, Object[]> rawDataMap) throws ParseException {
-    HashMap<String, Float> sma = m_SMAMap.get(smaDuration);
+    HashMap<String, Double> sma = m_SMAMap.get(smaDuration);
     if (sma == null) {
       m_SMAMap.put(smaDuration, new HashMap());
       sma = m_SMAMap.get(smaDuration);
@@ -79,7 +79,7 @@ public class PriceMovingAverage {
   }
 
   // Standard deviation, the bone of bollinger channel
-  public Float getSTD(String date, int distance, int stdDuration,
+  public Double getSTD(String date, int distance, int stdDuration,
           ConcurrentHashMap<String, Object[]> rawDataMap) throws ParseException {
     LocalDate start_date = MyUtils.getUsableDate(
             date, rawDataMap, distance, stdDuration, false, true);
@@ -89,8 +89,8 @@ public class PriceMovingAverage {
     String start_day = start_date.toString();
 
     Object[] raw_data = rawDataMap.get(start_day);
-    float p = (float) raw_data[YAHOO_DATA_INDEX.CLOSE.ordinal()];
-    HashMap<String, Float> std = m_STDMap.get(stdDuration);
+    double p = (double) raw_data[YAHOO_DATA_INDEX.CLOSE.ordinal()];
+    HashMap<String, Double> std = m_STDMap.get(stdDuration);
     if (std == null) {
       calculateSTD(stdDuration, rawDataMap);
       std = m_STDMap.get(stdDuration);
@@ -101,17 +101,17 @@ public class PriceMovingAverage {
     return std.get(start_day) / p;
   }
 
-  public Float getMACD(String date, int shortD, int longD, int midD,
+  public Double getMACD(String date, int shortD, int longD, int midD,
           ConcurrentHashMap<String, Object[]> rawDataMap) throws ParseException {
-    Float dea = getDEA(date, shortD, longD, midD, rawDataMap);
-    Float dif = getDIF(date, shortD, longD, rawDataMap);
+    Double dea = getDEA(date, shortD, longD, midD, rawDataMap);
+    Double dif = getDIF(date, shortD, longD, rawDataMap);
     if (dea == null || dif == null) {
       return null;
     }
     return (dif - dea) * 2;
   }
 
-  public Float getDEA(String date, int shortD, int longD, int midD,
+  public Double getDEA(String date, int shortD, int longD, int midD,
           ConcurrentHashMap<String, Object[]> rawDataMap) throws ParseException {
     LocalDate start_date = MyUtils.getUsableDate(date, rawDataMap, midD, midD, true, true);
     LocalDate end_date = MyUtils.getUsableDate(date, rawDataMap, midD, midD, false, true);
@@ -119,32 +119,32 @@ public class PriceMovingAverage {
       return null;
     }
     EMovingAverage ema = new EMovingAverage(2.0 / (1.0 + midD));
-    float v = 0;
+    double v = 0;
     while (start_date.isBefore(end_date)) {
-      Float dif = getDIF(start_date.toString(),
+      Double dif = getDIF(start_date.toString(),
               shortD, longD, rawDataMap);
       if (dif != null) {
-        v = (float) ema.average(dif);
+        v = (double) ema.average(dif);
       }
      start_date= start_date.plusDays(1);
     }
     return v;
   }
 
-  public Float getDIF(String date, int shortD, int longD,
+  public Double getDIF(String date, int shortD, int longD,
           ConcurrentHashMap<String, Object[]> rawDataMap) throws ParseException {
-    Float shortema = getRawEMA(date, shortD, rawDataMap);
-    Float longema = getRawEMA(date, longD, rawDataMap);
+    Double shortema = getRawEMA(date, shortD, rawDataMap);
+    Double longema = getRawEMA(date, longD, rawDataMap);
     if (shortema == null || longema == null) {
       return null;
     }
-    float dif = shortema - longema;
+    double dif = shortema - longema;
     return dif;
   }
 
   private void calculateEMA(int emaDuration,
           ConcurrentHashMap<String, Object[]> rawDataMap,
-          HashMap<String, Float> ema) throws ParseException {
+          HashMap<String, Double> ema) throws ParseException {
 
     LocalDate start_date = StatCalculator.getFirstValidDate(rawDataMap);
     Object[] raw_data = rawDataMap.get(start_date.toString());
@@ -162,7 +162,7 @@ public class PriceMovingAverage {
         buffer--;
       } else {
         buffer = 15;
-        ema.put(datestr, (float) ma.average((float) raw_data[YAHOO_DATA_INDEX.CLOSE.ordinal()]));
+        ema.put(datestr, (double) ma.average((double) raw_data[YAHOO_DATA_INDEX.CLOSE.ordinal()]));
       }
       start_date=start_date.plusDays(1);
     }
@@ -170,9 +170,9 @@ public class PriceMovingAverage {
 
   private void calculateSMA(int smaDuration,
           ConcurrentHashMap<String, Object[]> rawDataMap,
-          HashMap<String, Float> sma) throws ParseException {
+          HashMap<String, Double> sma) throws ParseException {
 
-    HashMap<String, Float> std = m_STDMap.get(smaDuration);
+    HashMap<String, Double> std = m_STDMap.get(smaDuration);
     if (std == null) {
       m_STDMap.put(smaDuration, new HashMap());
       std = m_STDMap.get(smaDuration);
@@ -197,14 +197,14 @@ public class PriceMovingAverage {
         buffer--;
       } else {
         buffer = 15;
-        ma.newNum((float) raw_data[YAHOO_DATA_INDEX.CLOSE.ordinal()]);
+        ma.newNum((double) raw_data[YAHOO_DATA_INDEX.CLOSE.ordinal()]);
         Object[] queue_data = ma.getWindow().toArray();
         double[] fqueue = new double[queue_data.length];
         for (int i = 0; i < queue_data.length; i++) {
           fqueue[i] = (double) queue_data[i];
         }
-        std.put(datestr, (float) deviation.evaluate(fqueue));
-        sma.put(datestr, (float) ma.getAvg());
+        std.put(datestr, (double) deviation.evaluate(fqueue));
+        sma.put(datestr, (double) ma.getAvg());
       }
      start_date= start_date.plusDays(1);
     }
